pyproject.toml
[tool.poetry]
name = "telegram-bot"
version = "2.2.0"
description = "Best way to create a scalable telegram bot with analytics"
homepage = "https://github.com/donBarbos/telegram-bot-template"
repository = "https://github.com/donBarbos/telegram-bot-template"
authors = ["donBarbos <donbarbos@proton.me>"]
license = "LGPL-v3.0"
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.10"
sqlalchemy = "^2.0.29"
pydantic = "^2.3.0"
pydantic-settings = "^2.0.3"
tinkoff-investments = "^0.2.0b97"
apscheduler = "^3.10.4"

[tool.poetry.group.bot.dependencies]
aiogram = "^3.3.0"
asyncpg = "^0.29.0"
redis = "^5.0.3"
loguru = "^0.7.2"
sentry-sdk = {extras = ["loguru"], version = "^1.45.0"}
aiohttp = {extras = ["speedups"], version = "^3.9.5"}
uvloop = "^0.19.0"
babel = "^2.13.1"
cachetools = "^5.3.2"
orjson = "^3.10.1"
alembic = "^1.13.1"
prometheus-client = "^0.20.0"

[tool.poetry.group.admin.dependencies]
gunicorn = "^22.0.0"
flask = "^3.0.3"
flask-admin = "^1.6.1"
flask-security-too = "^5.4.3"
flask-caching = "^2.1.0"
flask-babel = "^4.0.0"
flask-sqlalchemy = "^3.1.1"
psycopg2-binary = "^2.9.9"
tablib = {extras = ["xlsx"], version = "^3.6.1"}

[tool.poetry.group.dev.dependencies]
ruff = "^0.4.0"
mypy = "^1.7.1"
pre-commit = "^3.7.0"

[tool.ruff]
fix = true
unsafe-fixes = true
line-length = 120
select = ["ALL"]
ignore = ["D", "ANN101", "ANN102", "ANN401", "TD003", "FIX002", "COM812", "ISC001", "FBT001", "FBT002", "RUF", "ERA", "ARG005"]
cache-dir = "/tmp/ruff-cache/"
extend-exclude = ["logs", "configs", "migrations"]

[tool.ruff.isort]
no-lines-before = ["future", "standard-library"]
known-local-folder = ["bot", "admin"]

[tool.ruff.extend-per-file-ignores]
"tests/*.py" = ["ANN401", "S101", "S311"]

[tool.mypy]
python_version = "3.10"
files = "bot/*.py"
ignore_missing_imports = false
warn_return_any = true
warn_unused_configs = true
show_error_codes = true
show_error_context = true
pretty = true
disallow_subclassing_any = true
disallow_any_generics = true
disallow_untyped_calls = true
disallow_untyped_defs = true
disallow_incomplete_defs = true
check_untyped_defs = true
disallow_untyped_decorators = true
no_implicit_optional = true
warn_redundant_casts = true
warn_unused_ignores = true
follow_imports_for_stubs = true
namespace_packages = true
show_absolute_path = true

[[tool.mypy.overrides]]
module = "asyncpg.*"
ignore_missing_imports = true

[build-system]
requires = ["poetry-core>=1.0.0"]
build-backend = "poetry.core.masonry.api"
